Add files via upload
The console-based Hospital Management System is a Java-based application that enables users to manage patient and doctor information and schedule appointments. It allows adding patients and doctors, scheduling appointments with specific date and time, and generating patient, doctor, and appointment reports. The system efficiently tracks medical records and facilitates better organization and coordination in a hospital setting.